requirements.txt
# Backend dependencies
fastapi
uvicorn[standard]
pydantic
python-jose
passlib[bcrypt]
jwt
httpx
email-validator
python-multipart

# Data handling
kagglehub

# Testing
pytest
pytest-cov

# Development tools (optional)
flake8

bandit
pre-commit
pytest-mock==3.12.0

python-jose
